moved tabs to SettingsDialog removed Settings::query(), main window now deals with SettingsDialog directly
